+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 
 ## 1.0.0 (upcoming)
 
+* [GS-8334] Changes for stratio-identity feature
 * [GS-8313] Improve ldap identities management in BDT
 * [GS-8252] Create ldap groups using BDT
 * [GS-8186] Add step to get id from domain policy